# api/api-docs.yaml
openapi: 3.0.0
info:
  title: Greenalytic User Management API
  version: 1.0.0
  description: API for managing users with authentication and pagination
servers:
  - url: http://localhost:5000/greenalytic
    description: Local Development Server
  - url: https://greenalytic-vehicle-monitoring-1.onrender.com/greenalytic
    description: Production Server

# Define Authentication Scheme
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # JSON Web Token (JWT) format
...
paths:
...
  /emissions/vehicle/{vehicleId}:
    get:
      summary: Get emission data for a specific vehicle
      description: Retrieves emission data for a given vehicle with pagination, filtering by date range if provided.
      tags:
        - Emission Data
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema:
            type: integer
          description: The ID of the vehicle.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
          description: The page number for pagination (default is 1).
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
          description: The number of items per page (default is 10).
        - in: query
          name: startTime
          required: false
          schema:
            type: string
            format: date-time
          description: The start timestamp for filtering emission data (optional).
        - in: query
          name: endTime
          required: false
          schema:
            type: string
            format: date-time
          description: The end timestamp for filtering emission data (optional).
      responses:
        '200':
          description: A list of emission data with pagination metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The ID of the emission record.
                        vehicleId:
                          type: integer
                          description: The ID of the vehicle.
                        timestamp:
                          type: string
                          format: date-time
                          description: The timestamp when the emission data was recorded.
                        trackingDevice:
                          type: object
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                          description: The tracking device associated with the emission data.
                  meta:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                        description: The current page number.
                      totalPages:
                        type: integer
                        description: The total number of pages.
                      remainingItems:
                        type: integer
                        description: The number of items remaining after the current page.
                      totalItems:
                        type: integer
                        description: The total number of items matching the query.
                      limit:
                        type: integer
                        description: The number of items per page.
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message explaining what was wrong with the request.
        '500':
          description: Failed to fetch emission data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message explaining the failure.
    /vehicles/addvehicletouser/{userId}:
    post:
      summary: Add a Vehicle to a User
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plateNumber:
                  type: string
                  example: "ABC123"
                chassisNumber:
                  type: string
                  example: "CHS12345678"
                vehicleType:
                  type: string
                  example: "SUV"
                vehicleModel:
                  type: string
                  example: "Toyota Prado"
                yearOfManufacture:
                  type: integer
                  example: 2020
                usage:
                  type: string
                  example: "Personal"
      responses:
        "201":
          description: Vehicle added successfully
        "400":
          description: Bad request
        "500":
          description: Internal server error
  /vehicles/all:
    get:
      summary: Get All Vehicles (Paginated)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
          example: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          example: 10
      responses:
        "200":
          description: List of all vehicles
        "500":
          description: Internal server error
  /vehicles/{id}:
    get:
      summary: Get a Vehicle by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 3
      responses:
        "200":
          description: Vehicle details
        "404":
          description: Vehicle not found
        "500":
          description: Internal server error
  /vehicles/user/{userId}/vehicles:
    get:
      summary: Get Vehicles by User ID (Paginated)
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 2
        - name: page
          in: query
          required: false
          schema:
            type: integer
          example: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          example: 10
      responses:
        "200":
          description: List of vehicles for the user
        "500":
          description: Internal server error
  /vehicles/{vehicleId}/history:
    get:
      summary: Get Vehicle History
      security:
        - BearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
          example: 5
      responses:
        "200":
          description: Vehicle history retrieved successfully
        "404":
          description: Vehicle not found
        "500":
          description: Internal server error
  /vehicles/{vehicleId}/emissions/range:
    get:
      summary: Get Vehicle Emission Data in a Time Range
      security:
        - BearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
          example: 5
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-07T23:59:59Z"
      responses:
        "200":
          description: Emission data retrieved successfully
        "500":
          description: Internal server error
  /vehicles/{vehicleId}/fuels/range:
    get:
      summary: Get Vehicle Fuel Data in a Time Range
      security:
        - BearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
          example: 5
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-07T23:59:59Z"
      responses:
        "200":
          description: Fuel data retrieved successfully
        "500":
          description: Internal server error
  /vehicles/{vehicleId}/gps/range:
    get:
      summary: Get Vehicle GPS Data in a Time Range
      security:
        - BearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
          example: 5
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-07T23:59:59Z"
      responses:
        "200":
          description: GPS data retrieved successfully
        "500":
          description: Internal server error